Add tabs for Insights/Getting Started in home page
